Add DSF for address' region field.
Region used to require the full self_link of a region to be in the
config. Now it accepts the name or the self_link.

Note that this will have weird consequences if referenced from a
field that doesn't support both name and self_link as well, but that
honestly is a bug in the other field, and should be fixed there, in my
opinion.

The other option is a StateFunc that translates everything to a
self_link, at least making things consistent in state, but I wasn't sure
it was worth the effort. IF we like, I can take that route instead.