Fix commented out test.
Because of weirdness in our test runner, our CI tries to run every test
that matches the TestAcc* prefix, _even if it's commented out_. This
leads to failed tests in CI for tests that aren't even running, which is
annoying.

This PR just lowercases the first letter of the test so it doesn't match
the prefix anymore and the CI won't try to run it.