Add folder data source
It is common to split infrastructure in GCP by environments, where all are under the same
folder. You can think of it this way: you have a folder named "EnvA" that contains two
different projects, each project managed by a filesystem folder in the terraform
directory.

Currently, one would have to hardcode the folder name in the variables file (as an option)
after it had been created. With this data source it becomes possible to reference the
folder with a given display name (this has to been unique within the same parent).