Fix overly restrictive validation for ServiceAccountLinkRegex
A project name of the form `foo.io:foo` is perfectly legal, and a
service account for this project has the form

`projects/foo.io:foo/serviceAccounts/acctname@foo.foo.io.iam.gserviceaccount.com`

The two . characters in foo.foo.io fails the ProjectRegex validation,
making it impossible to create a google_service_account_key resource for
a service account in such a project.

This commit uses a looser validation for the project name the second
time it is used in the ServiceLinkAccountRegex.