Investigate and fix flaky sql user test.
This test fails for a few reasons: one, there are persistent quota
and service-unavailable errors with the second-gen SQL instances.
Two, the issue mentioned in #1184, that there's unspecified behavior
in tests where two resources have the same ID.  This changes the ID
to be user/host/instance instead of user/instance, and adds 429/503
handling to the sqladminOperationWait().