Fix panic conditions on TestAccContainerCluster_withAddons and other tests
The cluster name is is stored in a different key within the resource map
depending on if NodePool update is accessed from the cluster or node
pool resource.

The logic has been separated so construction of the NodePoolInformation
is now different depending on if the cluster or node pool uses it.