alphabatize/update sidebar (#1861)
I noticed we were missing the link for `region_disk`, and wanted to make sure we weren't missing any others. The easiest way to check was to alphabetize them, which I think is probably a good idea in general so I kept it. Turns out we were only missing `region_disk`, good job us.