revendor hashicorp/terraform (#1923)
`govendor fetch github.com/hashicorp/terraform/...@v0.11.8`

I figured this was a good time to do it since we just had a release. I verified that everything still compiles and that a few tests are passing- I don't expect anything else to change, but if there are any weird failures that this exposes we'll catch them in CI with plenty of time before the next release.