Revamp how IGM rollouts work.
When calling update, supply the entire IGM, not just the parts we want
to update, because this is an `update` call, not a `patch` call.

Retrieve the instance group manager before trying a rolling update, so
we can get the latest fingerprint, which is required for the update.

Catch a previously uncaught error when doing a rolling update, which was
hiding the failure message about the fingerprint from us.