Get rid of our gross state-inspection.
@rileykarson had a very good idea and also documentation handy that
showed that we don't actually need to detect whether the version is in
the state because it was auto-generated or because it's in the config.
Which is very cool. Documentation is here: https://cloud.google.com/compute/docs/instance-groups/updating-managed-instance-groups#relationship_between_versions_and_instancetemplate_properties_for_a_managed_instance_group

Relevant part is:

> You set both fields that do not match but only one value is different
> from the current instance template in the managed instance group.
>
> This is a valid request. The field that’s different from the current
> setting is taken as the intended value.

This didn't work for us originally because we were calling
SetInstanceTemplate, then calling performZoneUpdate, which would call
Update. At that point, the value the server had was the value in the
instance template, and so by the rule above, it would privilege the
template specified in version.

Except we don't need that call to SetInstanceTemplate if we're doing the
rolling update. So we moved it into the if block that doesn't do Update,
meaning the Update call would now be changing the setting from the value
in version to the value in instance_template, meaning the right field
would get privileged, and everything would just work, without having to
try and divine whether the version was set in the config or not.