Update our google.golang.org/api dependency.
Update all the packages under google.golang.org/api to use the same
commit. This is a necessary precursor for using Go modules.

There is a breaking API change in the clients, but it's a one line
change. As such, Travis will say this doesn't build. I've verified after
the one line change, it will run the unit tests. That change needs to be
made in MM, however. Once this PR is merged, a PR against MM will be
opened.