Add documentation on resource.google_container_cluster (#2419)
When a google_container_cluster is created using terraform the oauth_scope needs to be set to allow authentication of the VM service account on the google services. In addition, the default service account will need the IAM permissions for doing so.

The default service account will have this permissions if the project is created and the service is enabled using the Google Cloud Console. If those are done using terraform, the default account will not have this permissions.

I added some documentation regarding this issue, as it took me quite some time to figure out what was missing.

I hope, I placed this addition at the correct spot in the documentation.