Add gometalinter (#2627)
Parallel to https://github.com/terraform-providers/terraform-provider-azurerm/pull/2116

We need to run `gofmt -s` to switch over to the `gofmt` rule, and MM is going to complicate that. Once `1.20.0` is done, I'll run that command on flat MM files & add a `gofmt -s` step to MM generation (goimports does gofmt but doesn't do the s flag for silly reasons).

I'll duplicate this over to tpgb once this PR is merged.