Fix compute metadata error
This resolves the following issue:
 Error: Error refreshing state: 1 error(s) occurred:

* google_compute_project_metadata.default: 1 error(s) occurred:

2019-01-08T17:58:17.186+0100 [DEBUG] plugin.terraform-provider-gsuite:
2019/01/08 17:58:17 [ERR] plugin: plugin server: accept unix
/tmp/plugin123032284: use of closed network connection
* google_compute_project_metadata.default:
google_compute_project_metadata.default: Error reading Project metadata
for project "common_metadata": googleapi: Error 400: Invalid value
'common_metadata'. Values must match the following regular expression:
'(?:(?:[-a-z0-9]{1,63}\.)*(?:[a-z](?:[-a-z0-9]{0,61}[a-z0-9])?):)?(?:[0-9]{1,19}|(?:[a-z0-9](?:[-a-z0-9]{0,61}[a-z0-9])?))',
invalidParameter

This occurs when a resource exists that has "common_metadata" as the ID,
probably set by a previous provider.