Update go versions used in development (#2934)
Adding `.x` to Travis means use the latest minor version according to its docs.

I didn't update the "Requirements" header because this is just because of Go Modules; see https://github.com/terraform-providers/terraform-provider-google/pull/2932 for why. Since it's just a development change, it doesn't impact _using_ the provider.

Although, out of curiosity- is that header for development? Or is there a real requirement to have Go installed to use Terraform providers?