Update gometalinter for modules, add staticcheck (#2908)
This isn't working locally, but it matches the changes made in https://github.com/terraform-providers/terraform-provider-azurerm/pull/2722 so maybe something freaky is going on and it will only work in CI?

Edit: This only works in CI (well, at least doesn't work locally for *me*) but the same is true in the Azure provider. Most of those rules can be turned back on provided we make the corresponding code changes, but let's bring back `unused` for now. I'll dupe over to tpgb on approval.