Explain why scopes are required
Scopes are the legacy way to assign permissions
to an instance, so it was incredibly confusing
for a first time GCP user why this field was required.

This updates the terraform docs to explain this briefly
and link to the Google docs which state that this field
is still required  (presumably, for legacy reasons).