Make google_iam_policy page `NOTE` more prominent (#5638)
Because of the API the terraform provider uses (at the time of this
writing), defining and applying a google_iam_policy resource will
overwrite any existing policies for the given project. It may be
reasonable to assume that this resource could be additive, so moving
the `NOTE` higher on the page can help reduce issues caused by these
bad assumptions.