Allow `projects/` prefix for the project field in project IAM resources
This patch allows the user to specify the value of the project field for
project IAM resources as `project-id` or `projects/project-id`.

Note that the above was already true for google_project_iam_policy [1].
This patch ensures that the above is also true for other project IAM
resources like google_project_iam_member and google_project_iam_binding.

[1] See commit 2f2f19f2776d238fe854088f9906cff8b1a2db67