Allow passwords and tokens to be sensitive (#3177) (#5873)
* Allow passwords and tokens to be sensitive

Notification channels are configured by sending tokens and passwords via
labels. This change allows the sensitive labels to be configured in a
separate block so that they can be marked as Sensitive and hidden from plan
output. Encoders and Decoders are used to manipulate the labels block before
it is passed to the expanders/flatteners.

The token can be specified in either place right now, but the intent is to
remove the ability to manage these from being managed via the "labels" block
in the next breaking change.

* update description and remove old files

* Add sensitive label example

Signed-off-by: Modular Magician <magic-modules@google.com>