Updated docs for switching to/from beta
The "Resources can safely be converted from one version to the other without needing to rebuild infrastructure." statement and associated migration steps in the previous docs appears to be inaccurate as of Terraform v0.13.

Following the previously documented instructions for switching existing resources to use the google-beta provider resulted in `terraform plan` that attempted to destroy (using the google provider) all existing resources and to (re)create new instances of each resource (using the google-beta provider), instead of a no-op plan.

Also, if the existing provider's type was changed (rather than adding a new, separate provider, while retaining the old provider) in the Terraform configuration, then `terraform refresh` would also result in errors similar to the following:

```
Error: Provider configuration not present

To work with
google_compute_RESOURCE_TYPE.RESOURCE-NAME
its original provider configuration at
provider["registry.terraform.io/hashicorp/google"].PROVIDER-ALIAS
is required, but it has been removed. This occurs when a provider
configuration is removed while objects created by that provider still exist in
the state. Re-add the provider configuration to destroy
google_compute_RESOURCE_TYPE.RESOURCE-NAME,
after which you can remove the provider configuration again.
```

This new/undocumented behavior when switching between the google and google-beta providers appears to be related to changes in Terraform v0.12 and v0.13 that affected the way in which providers are referenced within the Terraform state data.