Refer to the value of `labels` as a 'map' (#3991) (#8045)
Labels is described as a 'set of key/value label pairs'. This is correct from an English grammar perspective, however in our case a Terraform 'set' is a strictly defined collection type, described here https://www.terraform.io/docs/configuration/types.html#set-

To avoid misleading users, can we refer to it as a 'map' instead, as this is the actual Terraform collection type we want.

Co-authored-by: Jak <44370243+jak-sdk@users.noreply.github.com>
Signed-off-by: Modular Magician <magic-modules@google.com>

Co-authored-by: Jak <44370243+jak-sdk@users.noreply.github.com>