Stopped marking provider as a /v3 library
This follows guidance from hashicorp to not import providers as go modules. See https://www.terraform.io/docs/extend/best-practices/depending-on-providers.html for more details.

We originally added /v3 as part of upgrading terraform-validator to the latest versions of the terraform provider; this inadvertently broke proper setting of the release version in the release pipeline.